Renamed module setup
