+++ python/grass/jupyter/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import absolute_import
 
-from .setup.funcs import *
+from .setup import *
